scripts/requirements.txt
aiohttp>=3.9
soundfile>=0.12
numpy>=1.24
openai>=1.0
google-cloud-speech>=2.0
pyyaml>=6.0
python-dotenv>=1.0
huggingface-hub>=0.20
